+++ src/main/java/de/retest/web/selenium/AutocheckingRecheckDriver.java
@@ -47,15 +47,15 @@ public class AutocheckingRecheckDriver extends UnbreakableDriver {
 	 *            The {@link RemoteWebDriver} to wrap.
 	 * @param options
 	 *            The {@link RecheckOptions} to use.
-	 * @param namingStrategy
+	 * @param checkNamingStrategy
 	 *            The {@link AutocheckingCheckNamingStrategy} to use.
 	 */
 	@Deprecated
 	public AutocheckingRecheckDriver( final RemoteWebDriver wrapped, final RecheckOptions options,
-			final AutocheckingCheckNamingStrategy namingStrategy ) {
+			final AutocheckingCheckNamingStrategy checkNamingStrategy ) {
 		super( wrapped );
 		this.options = options;
-		checkNamingStrategy = RecheckWebOptions.builder().build().getCheckNamingStrategy();
+		this.checkNamingStrategy = checkNamingStrategy;
 	}
 
 	public AutocheckingRecheckDriver( final RemoteWebDriver wrapped, final RecheckWebOptions options ) {
@@ -112,13 +112,13 @@ public class AutocheckingRecheckDriver extends UnbreakableDriver {
 	}
 
 	@Override
-	public WebElement findElement( final ByBestMatchToRetestId by ) {
+	public AutocheckingWebElement findElement( final ByBestMatchToRetestId by ) {
 		final WebElement wrapped = super.findElement( by );
 		String result = wrapped.toString();
 		// replace " -> xpath: HTML[1]/BODY[1]/A[1]"
 		result = result.substring( 0, result.lastIndexOf( " -> xpath: " ) );
 		final String representation = result + " -> retestId: " + by.getRetestId() + "]";
-		return new WebElementWrapper( wrapped, this ) {
+		return new AutocheckingWebElement( wrapped, this ) {
 			@Override
 			public String toString() {
 				return representation;
@@ -127,19 +127,19 @@ public class AutocheckingRecheckDriver extends UnbreakableDriver {
 	}
 
 	@Override
-	public WebElement findElement( final By by ) {
+	public AutocheckingWebElement findElement( final By by ) {
 		final WebElement result = super.findElement( by );
-		if ( result instanceof WebElementWrapper ) {
+		if ( result instanceof AutocheckingWebElement ) {
 			// Element was not found, so we already have it wrapped
-			return result;
+			return (AutocheckingWebElement) result;
 		}
-		return new WebElementWrapper( result, this );
+		return new AutocheckingWebElement( result, this );
 	}
 
 	@Override
 	public List<WebElement> findElements( final By by ) {
 		return super.findElements( by ).stream() //
-				.map( element -> new WebElementWrapper( element, this ) ) //
+				.map( element -> new AutocheckingWebElement( element, this ) ) //
 				.collect( Collectors.toList() );
 	}